requirements.txt
playwright>=1.40.0
dnspython>=2.4.0
requests>=2.31.0